+++ src/main/java/com/yaquodorg/yaquod/service/trip/TripServiceImpl.java
@@ -38,7 +38,7 @@ public class TripServiceImpl implements TripService {
     private static final String TOPIC_ETA_TRIP = "topic/trip/eta";
 
     @Override
-    public void initTrip(long requestId, long startLong, long startLat, long endLong, long endLat) {
+    public void initTrip(Long requestId, double startLong, double startLat, double endLong, double endLat) {
 
         // match vehicle
         List<Vehicle> vehicles = vehicleService.findKNearestVehicles(startLong, startLat, 1);
